+++ js/const.js
@@ -1,0 +1,1 @@
+export const MOCK_DATA_SIZE = 25;
